backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
motor>=3.3
pydantic-settings>=2.1
python-multipart>=0.0.7
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
Pillow>=10.0
orjson>=3.9
